+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Java packages start with eu.ill.cameo.
 * Corrected server responder and subscriber proxy ports to use loopback address.
 * Server configuration proxy ports can be overriden by command line.
+* Exports for Windows DLL.
 
 2.2.0
 -----